add base layout with color store provider
