+++ src/components/common/index.ts
@@ -1,5 +1,2 @@
-export { ThemeScript, ThemeSwitcher } from "./theme";
-export { SkeletonColor } from "./skeleton";
-
-export { default as DebugMode } from "./debug";
+export { default as BaseLayout } from "./base";
 export { default as Modal } from "./modal";
